requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
    # via kombu
appdirs==1.4.4
    # via fs
attrs==21.2.0
    # via pytest
    # via celery
bok-choy==1.1.1
    # via -r requirements/test.in
    # via
    #   -c requirements/constraints.txt
    #   event-tracking
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.6
    # via requests
    # via
    #   -c requirements/constraints.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    # via celery
click-plugins==1.1.1
    # via celery
    # via celery
code-annotations==1.2.0
    # via -r requirements/test.in
coverage==5.5
    # via pytest-cov
cryptography==3.4.8
    # via pyjwt
ddt==1.4.2
    # via -r requirements/test.in
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.in
    #   django-crum
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-i18n-tools
    #   edx-when
    #   event-tracking
    #   jsonfield2
    #   rest-condition
django-crum==0.7.9
    # via
    #   -r requirements/base.in
    #   edx-django-utils
django-ipware==4.0.0
    # via -r requirements/base.in
django-model-utils==4.1.1
    # via
    #   -r requirements/base.in
    #   edx-when
django-simple-history==3.0.0
    # via -r requirements/base.in
django-waffle==2.2.1
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
django-webpack-loader==0.7.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    # via
    #   -r requirements/base.in
    #   drf-jwt
    #   edx-drf-extensions
    #   rest-condition
drf-jwt==1.19.0
    # via
    #   -c requirements/common_constraints.txt
    #   edx-drf-extensions
edx-django-utils==4.4.0
    # via
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   edx-when
edx-drf-extensions==7.0.1
    # via
    #   -r requirements/base.in
    #   edx-when
edx-i18n-tools==0.7.0
    # via -r requirements/test.in
edx-opaque-keys==2.2.2
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-when
edx-rest-api-client==5.4.0
    # via -r requirements/base.in
edx-when==2.2.0
    # via -r requirements/base.in
event-tracking==1.1.1
    # via -r requirements/base.in
execnet==1.9.0
    # via pytest-xdist
freezegun==1.1.0
    # via -r requirements/test.in
fs==2.4.13
    # via xblock
future==0.18.2
    # via pyjwkest
httmock==1.4.0
    # via -r requirements/test.in
httpretty==1.1.4
    # via -r requirements/test.in
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via code-annotations
jsonfield2==3.0.3
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    # via celery
lazy==1.4
    # via bok-choy
logilab-common==1.8.1
    # via -r requirements/test.in
lxml==4.6.3
    # via xblock
markupsafe==2.0.1
    # via
    #   jinja2
    #   xblock
mock==4.0.3
    # via -r requirements/test.in
mypy-extensions==0.4.3
    # via logilab-common
newrelic==7.0.0.166
    # via edx-django-utils
packaging==21.0
    # via pytest
path==16.2.0
    # via edx-i18n-tools
pbr==5.6.0
    # via stevedore
pluggy==1.0.0
    # via pytest
polib==1.1.1
    # via edx-i18n-tools
    # via click-repl
psutil==5.8.0
    # via edx-django-utils
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pycparser==2.20
    # via cffi
pycryptodomex==3.10.1
    # via
    #   -r requirements/base.in
    #   pyjwkest
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==2.1.0
    # via
    #   -r requirements/base.in
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-rest-api-client
pymongo==3.12.0
    # via
    #   edx-opaque-keys
    #   event-tracking
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-xdist
pytest-cov==2.12.1
    # via -r requirements/test.in
pytest-django==4.4.0
    # via -r requirements/test.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-xdist==2.4.0
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   freezegun
    #   xblock
python-slugify==5.0.2
    # via code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.in
    #   celery
    #   django
    #   event-tracking
    #   fs
    #   xblock
pyyaml==5.4.1
    # via
    #   code-annotations
    #   edx-i18n-tools
    #   xblock
requests==2.26.0
    # via
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   httmock
    #   pyjwkest
    #   responses
    #   slumber
responses==0.14.0
    # via -r requirements/test.in
rest-condition==1.0.3
    # via edx-drf-extensions
rules==3.0
    # via -r requirements/base.in
selenium==3.141.0
    # via
    #   -r requirements/test.in
    #   bok-choy
semantic-version==2.8.5
    # via edx-drf-extensions
six==1.16.0
    # via
    #   bok-choy
    #   click-repl
    #   edx-drf-extensions
    #   edx-i18n-tools
    #   event-tracking
    #   fs
    #   pyjwkest
    #   python-dateutil
    #   responses
slumber==0.7.1
    # via edx-rest-api-client
sqlparse==0.4.2
    # via django
stevedore==3.4.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
sure==1.2.7
    # via -r requirements/test.in
testfixtures==6.18.2
    # via -r requirements/test.in
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
typing-extensions==3.10.0.2
    # via logilab-common
urllib3==1.26.7
    # via
    #   requests
    #   responses
    #   selenium
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
web-fragments==1.1.0
    # via xblock
webob==1.8.7
    # via xblock
xblock==1.5.1
    # via edx-when

# The following packages are considered to be unsafe in a requirements file:
# setuptools
